backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai==1.*
python-dotenv==1.0.1
pypdf==5.1.0

motor==3.7.0
beanie==1.27.0
reportlab==4.2.5


